% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.data.R
\name{clean.data}
\alias{clean.data}
\title{Cleans FaceReader data frame}
\usage{
clean.data(data, include = c("Basic", "All"))
}
\arguments{
\item{data}{The data frame created by \code{read.facereader}}

\item{include}{Value of either: \itemize{
\item \emph{Basic} for only the emotion values to be cleaned
\item \emph{All} for all values (columns) to be cleaned
}}
}
\value{
A cleaned data frame ready for further processing.
}
\description{
\code{clean.data} removes all \emph{"FIT_FAILED"} and \emph{"FIND_FAILED"} values and replaces them with \emph{NA}.
This function also appropriately changes each column value to the approrpriate type (e.g., numeric or factor). This is a
a good second step in the FaceReader data analysis pipeline, as the file must be cleaned before other commands such as
\code{score.aus} are performed on the data frame.
}
\examples{
data <- read.facereader("path/to/folder/")
data <- clean.data(data, include = "All")

}

