% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.x.id.R
\name{make.sub.id}
\alias{make.sub.id}
\title{Makes additional values for use in FaceReader output data frame}
\usage{
make.sub.id(data)
}
\arguments{
\item{data}{the data frame produced by \code{read.facereader}.}
}
\value{
A unique identifier for each participant, analysis, or condition based off of the naming structure of the files.
}
\description{
Functions to help parse out information from the naming structure of the files analyzed by FaceReader.
}
\details{
In order to use these functions, the FaceReader output files should be named in the following manner:
Participant#_Analysis#_extra_details.txt, where 'Participant#' is a unique numeric identifier, the first number
of which denotes the condition that participant was in. FaceReader should automatically provide an analysis number if
participants have more than one video being analyzed. Note, \code{make.sub.id} should still work if this naming convention
is not followed as long as each file being analyzed has a numeric value.
}
\examples{
data$id <-  make.sub.id(data)
data$analaysis <- make.analysis.id(data)
data$condition <- make.condition.id(data)

}

