% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{rename_aus}
\alias{rename_aus}
\title{Rename Action Units to their shortened names}
\usage{
rename_aus(data)
}
\arguments{
\item{data}{}
}
\value{
Dataframe with AUs renamed to shortened values.
}
\description{
Rename Action Units to their shortened names
}
