% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{score_aus}
\alias{score_aus}
\title{Score Action Units numerically}
\usage{
score_aus(data)
}
\arguments{
\item{data}{}
}
\value{
Dataframe with AUs recoded into numeric values.
}
\description{
Score Action Units numerically
}
