% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.facereader.R
\name{read.facereader}
\alias{read.facereader}
\title{Reads FaceReader output data}
\usage{
read.facereader(path, skip = 8)
}
\arguments{
\item{path}{The complete file path to location (folder) of your text files.}

\item{skip}{The number of lines in the text file to skip before the headers should be read. Defaults to 8 lines.}
}
\value{
A data frame that includes output provided by FaceReader 6.xx.
}
\description{
\code{read.facereader} reads in FaceReader 6.x text files, strips excess front matter,
and aggragates them into a single data frame for further analysis.
}
\details{
The file path (\code{path}) should lead to a directory that contains all of the text
files that will be read into the R workspace.
}
\examples{
data <- read.facereader("C:/My/Hard/Drive/Data/")

}

