% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{read_facereader}
\alias{read_facereader}
\title{Read Facereader file}
\usage{
read_facereader(file, subject = NULL, skip = NULL, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link[utils]{read.table}}}
}
\description{
Read Facereader file
}
\examples{

}
